+++ backend/internal/domain/repository/hello_world_repository.go
@@ -1,4 +1,4 @@
-//go:generate mockgen -source=hello_world_repository.go -destination=mock/hello_world_repository_mock.go
+//go:generate mockgen -source=hello_world_repository.go -destination=mock/hello_world_repository_mock.go -package mock
 package repository
 
 import (
