+++ backend/internal/domain/repository/mock/hello_world_repository_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=hello_world_repository.go -destination=mock/hello_world_repository_mock.go
+//	mockgen -source=hello_world_repository.go -destination=mock/hello_world_repository_mock.go -package mock
 //
 
-// Package mock_repository is a generated GoMock package.
-package mock_repository
+// Package mock is a generated GoMock package.
+package mock
 
 import (
 	context "context"
