+++ backend/internal/domain/repository/mock/post_repository_mock.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=post_repository.go -destination=mock/post_repository_mock.go
+//	mockgen -source=post_repository.go -destination=mock/post_repository_mock.go -package mock
 //
 
-// Package mock_repository is a generated GoMock package.
-package mock_repository
+// Package mock is a generated GoMock package.
+package mock
 
 import (
 	context "context"
